// src/app/models/usuario.interface.ts
export interface UsuarioInterface {
    idUsuario?: any | null | undefined;
    documentoUsuario?: string | null | undefined;
    idTipoDocumento?: string | null | undefined;
    nombreUsuario?: string | null | undefined;
    apellidoUsuario?: string | null | undefined;
    correoUsuario?: string | null | undefined;
    telefonoUsuario?: string | null | undefined;
    contrasenaUsuario?: string | null | undefined;
    idEstado?: any | null | undefined;
    idRol?: string | null | undefined;
}